Don't scrape Wikipedia pages if covered by Wikidata
Most of these earlier members now have Riksdagen IDs in Wikidata, so we
don't need to also fetch them from Wikipedia.

We still need the later ones for now, but we should be able to get rid
of them soon, as I've also added as many of those as I can find into
Wikidata.